requirements.txt
#appdirs==1.4.4
#asgiref==3.3.1
#attrs==20.3.0
#autobahn==20.7.1
#Automat==20.2.0
#cffi==1.14.4
#channels==3.0.2
#constantly==15.1.0
#consumers==0.6.1
#cryptography==3.2.1
#daphne==3.0.1
#distlib==0.3.1
#dj-database-url==0.5.0
#dj-static==0.0.6
#Django==3.1.3
#django-braces==1.14.0
django-ckedit==1.0S
django-ckeditor==6.0.0
django-js-asset==1.2.2
#django-model-utils==4.0.0
easy-thumbnails==2.7.1
#env==0.1.0
#filelock==3.0.12
#hyperlink==20.0.1
#idna==2.10
incremental==17.5.0
#kiwisolver==1.3.1
Pillow==8.0.1
#psycopg2==2.8.6
#pyasn1==0.4.8
#pyasn1-modules==0.2.8
#pycparser==2.20
#PyHamcrest==2.0.2
#pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-decouple==3.3
#pytz==2020.4
#service-identity==18.1.0
#six==1.15.0
sqlparse==0.4.1
static3==0.7.0
#style==1.1.6
Twisted
#txaio==20.4.1
#update==0.0.1
#virtualenv
#zope.interface==5.2.0
gevent==1.2.2
gevent-socketio==0.3.6
gevent-websocket==0.10.1